Restructure plugin hooks
